v1 phase 2: delete write API + entity_edits writes
Filesystem becomes the only write path. HTTP write endpoints, the
applyEdit chokepoint, and the entity_edits audit table are all gone.

Removed:
  - routes/inbox.ts (POST /:space/inbox, POST /:space/sources/from-url,
    PUT /:space/sources/*).
  - lib/inbox.ts (slugify, resolveInboxPath, ADD_SOURCE_EXTENSIONS, etc.).
  - lib/source-write-guards.ts (assertSourcePath, assertTextContent,
    sanitizeCaller).
  - lib/file-edits.ts (applyEdit, validateWikiHtmlDocument).
  - lib/edit-context.ts (in-process role attribution registry).
  - lib/href-rewrite.ts (agent-tool href rewriter).
  - lib/path-lock.ts (replaced by an inline 12-line Map<string, Promise>
    in extractors/runner.ts).
  - lib/remote-fetch.ts (HTTP fetch helper for /sources/from-url).
  - GET /:space/recent route + entity_edits read paths.
  - sync.ts no longer writes entity_edits on file change or delete.
  - extractors/runner.ts no longer registers edit-context attribution.
  - Test files: inbox-api e2e, sources-from-url e2e, inbox-lib unit,
    file-edits unit, href-rewrite unit. Two entity_edits assertions
    stripped from asset-indexing.test.ts; applyEdit test block stripped
    from binary-ingestion.test.ts.

Added (small):
  - lib/path.ts (~30 LOC): safeResolve extracted out of file-edits so
    the reader and space-scoped routes still have path-traversal
    protection. file-edits.ts was the only home for safeResolve, but
    the wider chunk it lived in was all write-API code.

Verified:
  - npm run typecheck: clean.
  - npm test: 267 unit tests pass.
  - npm run test:e2e: 46 tests pass (6 conditional skips).
  - npm run build: produces dist/.
